+++ storybook/src/components/accordion/accordion.core-template.ts
@@ -29,7 +29,7 @@ export const coreAccordion: ComponentImplementation<Accordion<TemplateResult>> =
             }) =>
               html`<dso-accordion-section
                 ?open=${open}
-                handle-title=${handleTitle}
+                .handleTitle=${handleTitle}
                 heading=${heading}
                 handle-url=${ifDefined(handleUrl)}
                 status-description=${ifDefined(statusDescription)}