+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Added
 * Skiplink: New component ([#2746](https://github.com/dso-toolkit/dso-toolkit/issues/2746))
+* Accordion: Renvooi in titel van accordion section toestaan ([#2857](https://github.com/dso-toolkit/dso-toolkit/issues/2857))
 
 ### Fixed
 * I18n: Werkt niet in Angular app ([#2880](https://github.com/dso-toolkit/dso-toolkit/issues/2880))